DRAW-334 refactor: 코드 리뷰 반영
