+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/MainWebSocketHandler.kt
@@ -2,7 +2,7 @@ package com.xorker.draw.websocket
 
 import com.xorker.draw.exception.XorkerException
 import com.xorker.draw.mafia.UserConnectionUseCase
-import com.xorker.draw.support.logging.logger
+import com.xorker.draw.mafia.WaitingQueueUseCase
 import com.xorker.draw.websocket.exception.WebSocketExceptionHandler
 import com.xorker.draw.websocket.log.WebSocketLogger
 import com.xorker.draw.websocket.message.request.WebSocketRequestParser
@@ -24,8 +24,6 @@ internal class MainWebSocketHandler(
     private val webSocketLogger: WebSocketLogger,
     private val userConnectionUseCase: UserConnectionUseCase,
 ) : TextWebSocketHandler() {
-
-    private val logger = logger()
 
     override fun afterConnectionEstablished(session: WebSocketSession) {
         webSocketLogger.handshake(session)
